+++ src/parse32.h
@@ -5,7 +5,7 @@
 #include "internal.h"
 
 int
-findLoadAddr32(const void *pheader, uint16_t phnum, unsigned int *load_addr) EJ_HIDDEN;
+findLoadAddr32(const void *pheader, uint32_t phnum, unsigned int *load_addr) EJ_HIDDEN;
 
 int
 findShdrs32(ejElfInfo *info, const struct ehdrParams *params) EJ_HIDDEN;
